Finished Backend and Initialized Frontend

+++ backend/controllers/product.controller.js
@@ -31,7 +31,11 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;    
-        
+    
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid product id" });
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: 'Product deleted'});
